+++ CHANGELOG.md
@@ -39,6 +39,8 @@
 - `UILoading`:
   - Updated `asDIVElement` buildDOM calls to pass `treeMap` and `setTreeMapRoot`.
 
+- `MasonryItem` and `_MasonryRenderItem`:
+  - Updated DOM generation calls to pass `treeMap` and `setTreeMapRoot`.
 ## 3.0.0-beta.19
 
 - dom_tools: ^3.0.0-beta.13